requirements.txt
anyio==4.4.0
ariadne==0.23.0
graphql-core==3.2.3
idna==3.7
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.1
